applied MVC structure
